# apps/wanaku-barn-backend/src/main/webui/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    SemanticActionSelection:
      type: object
      properties:
        actionId:
          type: string
          description: Eligible native Kamelet identifier
        sha256:
          type: string
          description: SHA-256 pin for the exact selected native Kamelet bytes; omitted
            on legacy selections
        label:
          type: string
          description: Fixed unique semantic label; no_match is reserved
        criteria:
          type: string
          description: Plain text criterion for selecting this action
        configuration:
          type: object
          additionalProperties:
            type: string
          description: Deployment parameter values; credentials use env:VARIABLE_NAME
            references
    SemanticExample:
      type: object
      properties:
        message:
          type: string
          description: Classification example input
        expectedLabel:
          type: string
          description: Expected action label or no_match
        expectedBlocked:
          type:
          - boolean
          - "null"
          description: Expect guard rejection instead of classification; absent means
            false
    SemanticExpert:
      type: object
      required:
      - id
      - name
      - bean
      - dependency
      properties:
        id:
          type: string
          description: Configured expert catalog identifier
          pattern: "[a-z][a-z0-9_-]{0,63}"
        name:
          type: string
          description: Configured expert display name
          pattern: \S
          maxLength: 120
        bean:
          type: string
          description: Named Camel expert instance configured externally by the deployment
          pattern: "[A-Za-z][A-Za-z0-9_]{0,63}"
        dependency:
          type: string
          description: Maven group:artifact:version implementation dependency
          pattern: "[A-Za-z0-9_.-]+:[A-Za-z0-9_.-]+:[A-Za-z0-9_.-]+"
        supportsConfidence:
          type: boolean
          description: Whether this contract exposes confidence controls
        operations:
          type: array
          items:
            $ref: "#/components/schemas/SemanticExpertOperation"
          description: Administrator declared native operations; omitted legacy entries
            expose choice
          maxItems: 32
    SemanticExpertOperation:
      type: object
      required:
      - name
      - inputTypes
      - resultType
      properties:
        name:
          type: string
          description: Native operation name
          pattern: "[A-Za-z][A-Za-z0-9_-]{0,63}"
        inputTypes:
          type: array
          items:
            type: string
            pattern: \S
            maxLength: 64
          description: "Supported input kinds, for example text"
          maxItems: 16
          minItems: 1
        resultType:
          type: string
          description: "Native result kind, for example choice or boolean"
          pattern: \S
          maxLength: 64
        resultMeaning:
          type: string
          description: Human-readable meaning of the result
          maxLength: 512
        parameterSchema:
          type: object
          additionalProperties: {}
          description: "Draft-07 JSON Schema for operation parameters, without external\
            \ references"
...
    SemanticGuard:
      type: object
      properties:
        expertId:
          type: string
        operation:
          type: string
        parameters:
          type: object
          additionalProperties: {}
        rejectWhen:
          type: boolean
          description: Boolean verdict that rejects the request
...
    SemanticPreviewRequest:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          description: Example invocation message to classify without action dispatch
          pattern: \S
          maxLength: 8192
...
    SemanticRouterDefinition:
      type: object
      properties:
        id:
          type: string
          description: "Persisted draft identifier, assigned by Barn"
        name:
          type: string
          description: "Router name; start with an ASCII letter, then use ASCII letters,\
            \ digits, hyphens, or underscores; no spaces; at most 120 characters;\
            \ empty for an incomplete draft"
          pattern: "^([A-Za-z][A-Za-z0-9_-]*)?$"
          maxLength: 120
        description:
          type: string
          description: Business purpose and public MCP tool description
        toolName:
          type: string
          description: "Public MCP tool name; start with an ASCII letter, then use\
            \ ASCII letters, digits, hyphens, or underscores; no spaces; at most 64\
            \ characters; empty for an incomplete draft"
          pattern: "^([A-Za-z][A-Za-z0-9_-]*)?$"
          maxLength: 64
        profile:
          type: string
          description: Supported action input and output profile
        expertId:
          type: string
          description: Configured expert catalog identifier
        guard:
          description: Optional Boolean guard evaluated before classification
          type:
          - object
          - "null"
          anyOf:
          - $ref: "#/components/schemas/SemanticGuard"
          - type: "null"
        semanticInput:
          type: string
          description: Invocation field supplied to native semantic evaluation
        instructions:
          type: string
          description: Plain text instructions for native choice evaluation
        noMatchCriteria:
          type: string
          description: Criterion for the reserved no_match label
        actions:
          type: array
          items:
            $ref: "#/components/schemas/SemanticActionSelection"
          description: Fixed eligible actions with deployment configuration and selection
            criteria
        examples:
          type: array
          items:
            $ref: "#/components/schemas/SemanticExample"
          description: Saved classification examples and expected labels
...
    WanakuError:
      type: object
      properties:
        message:
          type: string
...
    WanakuResponseListSemanticAction:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/SemanticAction"
    WanakuResponseListSemanticExpert:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/SemanticExpert"
    WanakuResponseListSemanticPublication:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/SemanticPublication"
    WanakuResponseListSemanticRouterDefinition:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/SemanticRouterDefinition"
...
    WanakuResponseMapStringString:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: object
          additionalProperties:
            type: string
    WanakuResponseSemanticExpert:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticExpert"
    WanakuResponseSemanticPreview:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticPreview"
    WanakuResponseSemanticPublication:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticPublication"
    WanakuResponseSemanticResolvedPublication:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticResolvedPublication"
    WanakuResponseSemanticRouterDefinition:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticRouterDefinition"
    WanakuResponseSemanticValidation:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/SemanticValidation"
...
    WanakuResponseVoid:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data: {}
paths:
...
  /api/v1/semantic-routers:
    get:
      summary: List saved router drafts
      description: List saved router drafts. Publication does not activate a runtime.
      tags:
      - Semantic routers
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListSemanticRouterDefinition"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
    post:
      summary: Save a router draft
      description: Save a router draft. Publication does not activate a runtime.
      tags:
      - Semantic routers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticRouterDefinition"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticRouterDefinition"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
  /api/v1/semantic-routers/actions:
    get:
      summary: List eligible curated Kamelet actions
      description: "List current eligible native Kamelets. With definitionId, include\
        \ exact historical revisions selected by that saved definition. The current\
        \ flag identifies current catalog selections."
      tags:
      - Semantic routers
      parameters:
      - name: definitionId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListSemanticAction"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
  /api/v1/semantic-routers/experts:
    get:
      summary: List configured expert instances
      description: List configured expert instances. Publication does not activate
        a runtime.
      tags:
      - Semantic routers
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListSemanticExpert"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
    post:
      summary: Create expert metadata
      description: Manage administrator metadata; deployment owns runtime bean configuration.
      tags:
      - Semantic routers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticExpert"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticExpert"
        "400":
          description: Invalid request
        "404":
          description: Expert not found
        "409":
          description: Duplicate identity or referenced expert
        "422":
          description: Invalid expert metadata
  /api/v1/semantic-routers/experts/{id}:
    put:
      summary: Update expert metadata
      description: Manage administrator metadata; deployment owns runtime bean configuration.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticExpert"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticExpert"
        "400":
          description: Invalid request
        "404":
          description: Expert not found
        "409":
          description: Duplicate identity or referenced expert
        "422":
          description: Invalid expert metadata
    delete:
      summary: Delete unreferenced expert metadata
      description: Manage administrator metadata; deployment owns runtime bean configuration.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseVoid"
        "400":
          description: Invalid request
        "404":
          description: Expert not found
        "409":
          description: Duplicate identity or referenced expert
        "422":
          description: Invalid expert metadata
    get:
      summary: Read expert metadata
      description: Manage administrator metadata; deployment owns runtime bean configuration.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticExpert"
        "400":
          description: Invalid request
        "404":
          description: Expert not found
        "409":
          description: Duplicate identity or referenced expert
        "422":
          description: Invalid expert metadata
  /api/v1/semantic-routers/files:
    post:
      summary: Inspect generated semantic router files
      description: "Generate all catalog files from a complete draft. The catalog\
        \ name is semantic-preview and the revision is preview. This operation does\
        \ not save a draft, publish a catalog, activate a runtime, or call an expert."
      tags:
      - Semantic routers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticRouterDefinition"
        required: true
      responses:
        "200":
          description: Generated file paths and contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseMapStringString"
        "400":
          description: Name or tool name is invalid
        "422":
          description: Definition is incomplete or structurally invalid
  /api/v1/semantic-routers/resolve:
    get:
      summary: Resolve a published semantic router by name
      description: Resolve an exact saved name to immutable publication pins and a
        published expert snapshot. Draft edits do not select a new publication. An
        explicit revision overrides the current pointer. Legacy records can omit the
        expert snapshot.
      tags:
      - Semantic routers
      parameters:
      - name: name
        in: query
        schema:
          type: string
          pattern: \S
          maxLength: 120
        required: true
      - name: revision
        in: query
        schema:
          type: string
          minLength: 1
          maxLength: 64
      responses:
        "200":
          description: Published selection resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticResolvedPublication"
        "400":
          description: Name or revision is invalid
        "404":
          description: "Router, publication, revision, or catalog not found"
        "409":
          description: Name is ambiguous or several legacy publications have no current
            selection
        "500":
          description: Stored publication and archive are inconsistent
  /api/v1/semantic-routers/validate:
    post:
      summary: Validate configuration without inference
      description: Validate configuration without inference. Publication does not
        activate a runtime.
      tags:
      - Semantic routers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticRouterDefinition"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticValidation"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
  /api/v1/semantic-routers/{id}:
    put:
      summary: Update a router draft
      description: Update a router draft. Publication does not activate a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticRouterDefinition"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticRouterDefinition"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
    delete:
      summary: Remove a router draft
      description: Remove a router draft. Publication does not activate a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseVoid"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
    get:
      summary: Read a router draft
      description: Read a router draft. Publication does not activate a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticRouterDefinition"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
  /api/v1/semantic-routers/{id}/preview:
    post:
      summary: Classify without executing actions
      description: Classify without executing actions. Publication does not activate
        a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticPreviewRequest"
        required: true
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticPreview"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
        "422":
          description: Definition is invalid
  /api/v1/semantic-routers/{id}/publish:
    post:
      summary: Publish an immutable catalog revision
      description: Publish an immutable catalog revision. Publication does not activate
        a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseSemanticPublication"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
        "422":
          description: Definition is invalid
  /api/v1/semantic-routers/{id}/revisions:
    get:
      summary: List published revisions
      description: List published revisions. Publication does not activate a runtime.
      tags:
      - Semantic routers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListSemanticPublication"
        "400":
          description: Invalid request
        "404":
          description: Definition not found
...
info:
  title: wanaku-barn-backend API
  version: 0.3.0-SNAPSHOT
servers:
- url: http://localhost:8180
  description: Auto generated value
- url: http://0.0.0.0:8180
  description: Auto generated value